+++ src/main/kotlin/com/github/fmueller/jarvis/conversation/Conversation.kt
@@ -53,6 +53,7 @@ data class Message(
             - ```/plain``` - Sends a chat message without code context
             - ```/copy``` - Copies the conversation to the clipboard
             - ```/model <modelName>``` - Changes the model to use (`default` is `qwen3:1.7b`)
+            - ```/model``` or ```/model-info``` - Shows the info card of the current model
             - ```/host <host>``` - Sets the Ollama host (`default` is `http://localhost:11434`)
             """.trimIndent()
         )